requirements.txt
black
isort
jinja2
markdown2
premailer
bs4
fastapi
uvicorn
websockets
matplotlib
pyyaml
